remove tf.cast

+++ attention.py
@@ -47,9 +47,9 @@ class BahdanauAttention(tf.keras.Model):
         self.W2 = tf.keras.layers.Dense(units)
         self.V = tf.keras.layers.Dense(1)
         self.pos = tf.zeros([batch_sz], dtype=tf.int32, name='attention_pos')
-        self.att_len = tf.cast(att_len, dtype=tf.int32)
+        self.att_len = att_len
         self.batch_sz = batch_sz
-        self.units = tf.cast(units, dtype=tf.int32)
+        self.units = units
 
     def fetch_att_values(self, values, pos, batch_sz, units_num):
         def cond(i, att_value, xpos):
